requirements.txt
Click==7.0
matplotlib==3.1.0
numpy==1.16.4
pandas==0.24.2
Pillow==6.1.0
scikit-learn==0.21.2
scipy==1.3.0
seaborn==0.9.0
torch==1.5.0
torchvision==0.6.0
